refactor: rename ThreeExtented -> ThreeExtended and MathExtented -> MathExtended
