+++ src/Core/Commander/Providers/WMS_Provider.js
@@ -11,7 +11,7 @@ import Fetcher from 'Core/Commander/Providers/Fetcher';
 import * as THREE from 'three';
 import Projection from 'Core/Geographic/Projection';
 import CacheRessource from 'Core/Commander/Providers/CacheRessource';
-import mE from 'Core/Math/MathExtented';
+import mE from 'Core/Math/MathExtended';
 import BoundingBox from 'Scene/BoundingBox';
 import { UNIT } from 'Core/Geographic/GeoCoordinate';
