+++ src/Core/Math/MathExtended.js
