+++ src/Globe/Globe.js
@@ -7,7 +7,7 @@
 import Layer from 'Scene/Layer';
 import Quadtree from 'Scene/Quadtree';
 import SchemeTile from 'Scene/SchemeTile';
-import MathExt from 'Core/Math/MathExtented';
+import MathExt from 'Core/Math/MathExtended';
 import TileMesh from 'Globe/TileMesh';
 import Atmosphere from 'Globe/Atmosphere';
 import Clouds from 'Globe/Clouds';
