+++ src/Globe/TileMesh.js
@@ -20,8 +20,8 @@ import NodeMesh from 'Renderer/NodeMesh';
 import TileGeometry from 'Globe/TileGeometry';
 import BoundingBox from 'Scene/BoundingBox';
 import * as THREE from 'three';
-import OBBHelper from 'Renderer/ThreeExtented/OBBHelper';
-import SphereHelper from 'Renderer/ThreeExtented/SphereHelper';
+import OBBHelper from 'Renderer/ThreeExtended/OBBHelper';
+import SphereHelper from 'Renderer/ThreeExtended/SphereHelper';
 import LayeredMaterial, { l_ELEVATION } from 'Renderer/LayeredMaterial';
 import GlobeDepthMaterial from 'Renderer/GlobeDepthMaterial';
 import MatteIdsMaterial from 'Renderer/MatteIdsMaterial';
