+++ src/Renderer/ThreeExtended/SphereHelper.js
