+++ src/Renderer/ThreeExtended/StarGeometry.js
